refactor: Flatten hooks folder
