+++ packages/react-responsive-pagination/src/hooks/useAvailableWidth.ts
