+++ packages/react-responsive-pagination/src/hooks/useIsUnmounted.ts
