+++ packages/react-responsive-pagination/src/hooks/useResizeNotifier.ts
