+++ packages/react-responsive-pagination/src/widthCalculator/numberWidthCalculator.ts
