+++ packages/react-responsive-pagination/src/widthCalculator/pageWidthCalculator.ts
